Make extern crate linked_hash_map conditional
The dependency on linked_hash_map is optional, but there was an extern crate linked_hash_map that was not under a conditional flag and this gives compiler errors, when "preserve_order" is not set since the dependency is missing.
This adds the conditional avoiding trying to import the crate when "preserve_order" is not set.